added camerazones
