+++ Assets/Scripts/Camera/CameraZoneTrigger.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 587dce3860a986a4a9c07f09f7bc31f2
