--- Assets/Scripts/Miscellaneous.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9b5d8d2dbbf830643ae2c95806dd202b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
