--- Assets/Scripts/Miscellaneous/CameraFollow.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 0cb9880862fd1d746b56ae3d838b0727
