--- Assets/Scripts/Miscellaneous/MenuController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: f8bed6357381541aa8446f43b9c76914
